requirements.txt
streamlit
openai
requests
beautifulsoup4
pypdf2
python-pptx
openpyxl
pandas